Clean up windowing functions
This commit splits out a bunch of functions that really don't need to be
sitting around in main.cpp taking up space. This commit creates a
windowing translation unit which is intended to handle the generally
arcane and annoying windowing logic of imgui, win32, and directx.
